Bank plugin fixes

+++ runelite-client/src/main/java/net/runelite/client/plugins/itemprices/ItemPricesOverlay.java
@@ -216,7 +216,7 @@ class ItemPricesOverlay extends Overlay
 
 		if (config.showGEPrice())
 		{
-			gePrice = itemManager.getItemPrice(id);
+			gePrice = (int) itemManager.getItemPrice(id);
 		}
 		if (config.showHAValue())
 		{
@@ -291,7 +291,7 @@ class ItemPricesOverlay extends Overlay
 
 	private int calculateHAProfit(int haPrice, int gePrice)
 	{
-		int natureRunePrice = itemManager.getItemPrice(ItemID.NATURE_RUNE);
+		int natureRunePrice = (int) itemManager.getItemPrice(ItemID.NATURE_RUNE);
 		return haPrice - gePrice - natureRunePrice;
 	}
